content/svg/content/test/test_pointer-events.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500174
-->
<head>
  <title>Test Pointer Events</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function run()
{
  var svgDoc = document.getElementById('svg');
  var div = document.getElementById("div");
  var x = div.offsetLeft;
  var y = div.offsetTop;
  var good = document.getElementById('good');
  var fo = document.getElementById('fo');
  var elementFromPoint = document.elementFromPoint(55 + x, 55 + y);
  is(good, elementFromPoint, 'pointer-events="all"');
  elementFromPoint = document.elementFromPoint(205 + x, 55 + y);
  is(good, elementFromPoint, 'foreignObject with clip-path');
  elementFromPoint = document.elementFromPoint(205 + x + 20, 55 + y + 20);
  is(fo, elementFromPoint, 'foreignObject with clip-path');

  SimpleTest.finish();
}

]]>
</script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=500174">Mozilla Bug 500174</a>
<p id="display"></p>
<div id="content">

  <div width="100%" height="1" id="div">
  </div>
  <svg xmlns="http://www.w3.org/2000/svg" width="100%" height="100%" id="svg">
    <defs>
      <clipPath id="clip">
        <rect x="20" y="20" width="50" height="50"/>
      </clipPath>
    </defs>
    <rect id="bad" width="100%" height="100%" fill="green"/>
    <circle id="good" cx="50%" cy="50%" r="1" stroke-width="1000" fill="black" pointer-events="all"/>
      <foreignObject id="fo" x="200" y="50" width="50" height="50" clip-path="url(#clip)">
      <svg onload="run()">
        <rect width="100%" height="100%" fill="green"/>
      </svg>
    </foreignObject>
  </svg>

</div>
<pre id="test">
</pre>
</body>
</html>
